+++ apps/geoportal/src/app/config/layerconf.ts
@@ -1,0 +1,88 @@
+export const defaultLayerConf = {
+  namedStyles: {
+    default: { opacity: 0.6 },
+    night: {
+      opacity: 0.9,
+      "css-filter": "filter:grayscale(0.9)brightness(0.9)invert(1)",
+    },
+    blue: {
+      opacity: 1.0,
+      "css-filter":
+        "filter:sepia(0.5) hue-rotate(155deg) contrast(0.9) opacity(0.9) invert(0)",
+    },
+  },
+  defaults: {
+    wms: {
+      format: "image/png",
+      tiled: "true",
+      maxZoom: 22,
+      opacity: 0.6,
+      version: "1.1.1",
+      pane: "backgroundLayers",
+    },
+    vector: {},
+  },
+  namedLayers: {
+    "wupp-plan-live": {
+      type: "wms",
+      url: "https://geodaten.metropoleruhr.de/spw2/service",
+      layers: "spw2_light",
+      tiled: "false",
+      version: "1.3.0",
+    },
+    trueOrtho2020: {
+      type: "wms",
+      url: "https://maps.wuppertal.de/karten",
+      layers: "R102:trueortho2020",
+      transparent: true,
+    },
+    rvrGrundriss: {
+      type: "wmts",
+      url: "https://geodaten.metropoleruhr.de/spw2/service",
+      layers: "spw2_light_grundriss",
+      version: "1.3.0",
+      transparent: true,
+      tiled: false,
+    },
+    trueOrtho2022: {
+      type: "wms",
+      url: "https://maps.wuppertal.de/karten",
+      layers: "R102:trueortho2022",
+      transparent: true,
+    },
+    rvrSchriftNT: {
+      type: "wmts-nt",
+      url: "https://geodaten.metropoleruhr.de/dop/dop_overlay?language=ger",
+      layers: "dop_overlay",
+      version: "1.3.0",
+      tiled: false,
+      transparent: true,
+      buffer: 50,
+    },
+    rvrSchrift: {
+      type: "wmts",
+      url: "https://geodaten.metropoleruhr.de/dop/dop_overlay?language=ger",
+      layers: "dop_overlay",
+      version: "1.3.0",
+      tiled: false,
+      transparent: true,
+    },
+    amtlich: {
+      type: "tiles",
+      maxNativeZoom: 20,
+      maxZoom: 22,
+      url: "https://geodaten.metropoleruhr.de/spw2?SERVICE=WMTS&REQUEST=GetTile&VERSION=1.0.0&LAYER=spw2_light&STYLE=default&FORMAT=image/png&TILEMATRIXSET=webmercator_hq&TILEMATRIX={z}&TILEROW={y}&TILECOL={x}",
+    },
+    basemap_relief: {
+      type: "vector",
+      style:
+        "https://sgx.geodatenzentrum.de/gdz_basemapde_vektor/styles/bm_web_top.json",
+    },
+    amtlichBasiskarte: {
+      type: "wmts",
+      url: "https://maps.wuppertal.de/karten",
+      layers: "abkf",
+      transparent: true,
+    },
+  },
+};
