--- libraries/appframeworks/portals/src/lib/carma-apps-portals.spec.ts
@@ -1,7 +1,0 @@
-import { carmaAppsPortals } from './carma-apps-portals';
-
-describe('carmaAppsPortals', () => {
-  it('should work', () => {
-    expect(carmaAppsPortals()).toEqual('carma-apps-portals');
-  });
-});
