+++ libraries/appframeworks/portals/src/lib/components/map-measure/InfoBoxMeasurement.jsx
