+++ libraries/appframeworks/portals/src/lib/components/map-measure/hexagon.png
