+++ libraries/appframeworks/portals/src/lib/components/map-measure/measure.png
