+++ libraries/appframeworks/portals/src/lib/components/map-measure/polygon-active.png
