+++ libraries/appframeworks/portals/src/lib/utils/helper.ts
@@ -1,51 +1,55 @@
-import { md5FetchText } from 'react-cismap/tools/fetching';
-import { getGazDataForTopicIds } from 'react-cismap/tools/gazetteerHelper';
-import clsx from 'clsx';
-import { twMerge } from 'tailwind-merge';
+import { md5FetchText } from "react-cismap/tools/fetching";
+import { getGazDataForTopicIds } from "react-cismap/tools/gazetteerHelper";
+import clsx from "clsx";
+import { twMerge } from "tailwind-merge";
+import { ENDPOINT } from "../types";
 
-export const getGazData = async (setGazData, host) => {
-  const prefix = 'GazData';
-  const sources: any = {};
-
-  sources.adressen = await md5FetchText(
-    prefix,
-    host + '/data/3857/adressen.json'
-  );
-  sources.bezirke = await md5FetchText(
-    prefix,
-    host + '/data/3857/bezirke.json'
-  );
-  sources.quartiere = await md5FetchText(
-    prefix,
-    host + '/data/3857/quartiere.json'
-  );
-  sources.pois = await md5FetchText(prefix, host + '/data/3857/pois.json');
-  sources.kitas = await md5FetchText(prefix, host + '/data/3857/kitas.json');
-
-  const gazData = getGazDataForTopicIds(sources, [
-    'pois',
-    'kitas',
-    'bezirke',
-    'quartiere',
-    'adressen',
-  ]);
-
-  setGazData(gazData);
+const buildHostUri = (host: string, endpoint: ENDPOINT, crs: string) => {
+  return `${host}/data/${crs}/${endpoint}.json`;
 };
 
-export function cn(...inputs) {
+export const getGazData = async (
+  host: string,
+  {
+    crs = "3857",
+    prefix = "GazData",
+    endpoints = [
+      ENDPOINT.ADRESSEN,
+      ENDPOINT.BEZIRKE,
+      ENDPOINT.QUARTIERE,
+      ENDPOINT.POIS,
+      ENDPOINT.KITAS,
+    ],
+  }: {
+    crs?: string;
+    prefix?: string;
+    endpoints?: ENDPOINT[];
+  } = {},
+) => {
+  const sources: Record<ENDPOINT, string> = Object.fromEntries(
+    await Promise.all(
+      endpoints.map(async (endpoint) => [
+        endpoint,
+        await md5FetchText(prefix, buildHostUri(host, endpoint, crs)),
+      ]),
+    ),
+  );
+  return getGazDataForTopicIds(sources, endpoints);
+};
+
+export function cn(...inputs: string[]) {
   return twMerge(clsx(inputs));
 }
 
 export const parseDescription = (description: string) => {
-  const result = { inhalt: '', sichtbarkeit: '', nutzung: '' };
-  const keywords = ['Inhalt:', 'Sichtbarkeit:', 'Nutzung:'];
+  const result = { inhalt: "", sichtbarkeit: "", nutzung: "" };
+  const keywords = ["Inhalt:", "Sichtbarkeit:", "Nutzung:"];
 
   if (!description) {
     return result;
   }
 
-  function extractTextAfterKeyword(input, keyword) {
+  function extractTextAfterKeyword(input: string, keyword: string) {
     const index = input.indexOf(keyword);
     if (index !== -1) {
       const startIndex = index + keyword.length;
@@ -58,18 +62,18 @@ export const parseDescription = (description: string) => {
       }
       return input.slice(startIndex, endIndex).trim();
     }
-    return '';
+    return "";
   }
 
-  result.inhalt = extractTextAfterKeyword(description, 'Inhalt:');
-  result.sichtbarkeit = extractTextAfterKeyword(description, 'Sichtbarkeit:');
-  result.nutzung = extractTextAfterKeyword(description, 'Nutzung:');
+  result.inhalt = extractTextAfterKeyword(description, "Inhalt:");
+  result.sichtbarkeit = extractTextAfterKeyword(description, "Sichtbarkeit:");
+  result.nutzung = extractTextAfterKeyword(description, "Nutzung:");
 
   return result;
 };
 
-export function paramsToObject(entries) {
-  const result = {};
+export function paramsToObject(entries: URLSearchParams) {
+  const result: { [key: string]: string } = {};
   for (const [key, value] of entries) {
     // each 'entry' is a [key, value] tupple
     result[key] = value;
