+++ libraries/appframeworks/portals/tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "extends": "../../../tsconfig.strict.base.json",
+  "extends": "../../../tsconfig.legacy.base.json",
   "compilerOptions": {},
   "files": [],
   "references": [
