+++ package.json
@@ -155,7 +155,7 @@
     "tailwindcss": "3.4.7",
     "ts-node": "^10.9.2",
     "typescript": "^5.5.4",
-    "typescript-eslint": "^7.18.0",
+    "typescript-eslint": "^8.0.0",
     "vite": "^5.3.5",
     "vite-plugin-comlink": "^5.0.1",
     "vite-plugin-dts": "^3.9.1",
